+++ scripts/macros/classFeatures/artificer/infusions/armorOfMagicalStrength.js
@@ -1,0 +1,4 @@
+import {chris} from '../../../../helperFunctions.js';
+export async function armorOfMagicalStrength({speaker, actor, token, character, item, args, scope, workflow}) {
+    await chris.removeCondition(workflow.actor, 'Prone');
+}